requirements.txt
fastapi===0.115.0
uvicorn==0.23.0
PyPDF2==3.0.1
spacy==3.8.4
scikit-learn==1.6.1
python-multipart
sentence-transformers
numpy
nltk
requests
langchain==0.3.18
langchain-community
langchain_openai
openai
beautifulsoup4==4.12.3
playwright==1.42.0
aiohttp==3.9.3
python-dotenv==1.0.1
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.3.0
pytest==7.3.1
pytest-asyncio
newrelic>=8.0.0